nlink: GraphSat 作り直し
