+++ programs/nlink/src/graphsat/GsEdge.h
