+++ programs/nlink/src/graphsat/GsNode.h
